refactor(#step2): 메뉴 및 메뉴 그룹 패키징 분리
