+++ src/main/java/kitchenpos/menus/application/MenuService.java
@@ -5,7 +5,9 @@ import kitchenpos.menus.dto.MenuCreateRequest;
 import kitchenpos.menus.dto.MenuDetailResponse;
 import kitchenpos.menus.dto.MenuProductElement;
 import kitchenpos.menus.mapper.MenuMapper;
-import kitchenpos.menus.tobe.domain.*;
+import kitchenpos.menus.tobe.domain.menu.*;
+import kitchenpos.menus.tobe.domain.menugroup.MenuGroup;
+import kitchenpos.menus.tobe.domain.menugroup.MenuGroupRepository;
 import kitchenpos.products.tobe.domain.Product;
 import kitchenpos.products.tobe.domain.ProductRepository;
 import kitchenpos.support.infra.PurgomalumClient;
