+++ src/main/java/kitchenpos/menus/tobe/domain/menu/Menu.java
@@ -1,4 +1,6 @@
-package kitchenpos.menus.tobe.domain;
+package kitchenpos.menus.tobe.domain.menu;
+
+import kitchenpos.menus.tobe.domain.menugroup.MenuGroup;
 
 import javax.persistence.*;
 import java.math.BigDecimal;
