+++ src/main/java/kitchenpos/menus/tobe/domain/menu/MenuName.java
@@ -1,4 +1,4 @@
-package kitchenpos.menus.tobe.domain;
+package kitchenpos.menus.tobe.domain.menu;
 
 import kitchenpos.support.infra.PurgomalumClient;
