+++ src/main/java/kitchenpos/menus/tobe/domain/menugroup/MenuGroupName.java
@@ -1,4 +1,4 @@
-package kitchenpos.menus.tobe.domain;
+package kitchenpos.menus.tobe.domain.menugroup;
 
 import javax.persistence.Column;
 import javax.persistence.Embeddable;
