+++ src/test/java/kitchenpos/menus/application/MenuServiceTest.java
@@ -5,7 +5,9 @@ import kitchenpos.menus.dto.MenuCreateRequest;
 import kitchenpos.menus.dto.MenuDetailResponse;
 import kitchenpos.menus.dto.MenuProductElement;
 import kitchenpos.menus.mapper.MenuMapper;
-import kitchenpos.menus.tobe.domain.*;
+import kitchenpos.menus.tobe.domain.menu.*;
+import kitchenpos.menus.tobe.domain.menugroup.MenuGroup;
+import kitchenpos.menus.tobe.domain.menugroup.MenuGroupRepository;
 import kitchenpos.products.infra.FakePurgomalumClient;
 import kitchenpos.products.tobe.domain.InMemoryProductRepository;
 import kitchenpos.products.tobe.domain.Product;
@@ -26,7 +28,7 @@ import java.util.*;
 import static java.util.Objects.isNull;
 import static java.util.stream.Collectors.toUnmodifiableList;
 import static kitchenpos.Fixtures.*;
-import static kitchenpos.menus.tobe.domain.FakeMenu.createFake;
+import static kitchenpos.menus.tobe.domain.menu.FakeMenu.createFake;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertAll;
