+++ src/test/java/kitchenpos/menus/tobe/domain/menugroup/FakeMenuGroup.java
@@ -1,4 +1,4 @@
-package kitchenpos.menus.tobe.domain;
+package kitchenpos.menus.tobe.domain.menugroup;
 
 public class FakeMenuGroup extends MenuGroup {
