+++ src/test/java/kitchenpos/menus/tobe/domain/menugroup/MenuGroupTest.java
@@ -1,4 +1,4 @@
-package kitchenpos.menus.tobe.domain;
+package kitchenpos.menus.tobe.domain.menugroup;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
